new: Add more routing, rename some files
