+++ src/components/favorite-page/favorites-page.jsx
