+++ src/components/main-page/main-page.jsx
