+++ src/components/reviews-list/reviews-list.jsx
@@ -7,7 +7,7 @@ const ReviewsList = ({reviews}) => {
     <section className="property__reviews reviews">
       <h2 className="reviews__title">Reviews &middot; <span className="reviews__amount">{reviews.length}</span></h2>
       <ul className="reviews__list">
-        {reviews.map((review) => (
+        {reviews.slice(0, 10).map((review) => (
           <li key={review.id} className="reviews__item">
             <div className="reviews__user user">
               <div className="reviews__avatar-wrapper user__avatar-wrapper">
