+++ src/index.js
@@ -8,7 +8,7 @@ import {getRandomIntegerNumber} from './utils/common.js';
 
 const isLogged = Math.random() > 0.5;
 const offers = generateOffers(20);
-const reviews = generateReviews(getRandomIntegerNumber(0, 7));
+const reviews = generateReviews(getRandomIntegerNumber(0, 15));
 const host = generateHost(10);
 ReactDOM.render(
     <App offers={offers} reviews={reviews} host={host} isLogged={isLogged} />,
